Rename file with has_ checks
