+++ shellwhat/checks/has_funcs.py
